mencoba merubah ke int karena sebelumnya error

+++ fungsitask.cpp
@@ -20,7 +20,8 @@ void output()
 {
     cout << "Hasilnya : " << LuasLingkaran(Phi, X, r) << endl;
 }
-float main()
+// Program Utama
+int main()
 {
     input();
     output();
